src/requirements.txt
pandas==1.3.3
pyyaml==5.4.1
torch==1.9.0